requirements.txt
keyboard==0.13.5
matplotlib==3.9.3
mss==10.0.0
numpy==2.1.3
Pillow==11.0.0
psutil==6.1.0
pyautogui==0.9.54
pydivert==2.1.0
pygetwindow==0.0.9
pyserial==3.5
pytesseract==0.3.13
screeninfo==0.8.1
win32gui==221.6

opencv-python~=4.7.0.72
cffi~=1.15.1
pip~=23.1.2
attrs~=23.1.0
google~=3.0.0
jinja2~=3.1.4
pypiwin32~=223
pywin32~=310
setuptools~=65.5.0
pyinstaller~=5.11.0
six~=1.16.0
pynput~=1.7.6
wcwidth~=0.2.6
PyMsgBox~=1.0.9
pyperclip~=1.8.2
MouseInfo~=0.1.3
PyScreeze~=0.1.28
pytweening~=1.0.7
packaging~=23.1
pyparsing~=3.2.0
prettytable~=3.7.0
PyRect~=0.2.0
easyocr~=1.7.2
Camelot~=12.6.29
tabula~=1.0.5
requests~=2.32.4
interception~=0.6.0
pathfinding~=1.0.17
pymem~=1.14.0
flask~=3.1.0
frida~=17.2.11
discord~=2.3.2
mitmproxy~=11.0.2
cryptography~=44.0.3